requirements.txt
python-telegram-bot==20.7
Django==4.2.7
django-environ==0.11.2
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-results==2.5.1
dj_database_url
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
